+++ workspaces/bulk-import/plugins/bulk-import/src/nfsExports.test.ts
@@ -1,0 +1,30 @@
+/*
+ * Copyright Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import translationsModuleDefault from './bulkImportTranslationsModuleExport';
+import { bulkImportTranslationsModule } from './index';
+
+describe('bulk-import NFS exports', () => {
+  it('should export a translations module as a FrontendModule', () => {
+    expect(bulkImportTranslationsModule).toBeDefined();
+    expect(bulkImportTranslationsModule.$$type).toBe(
+      '@backstage/FrontendModule',
+    );
+  });
+
+  it('should export the translations module as default for NFS discovery', () => {
+    expect(translationsModuleDefault).toBe(bulkImportTranslationsModule);
+  });
+});